Success React router-dom & SPA
