--- Lab_week_07/src/pages/DashboardPage.jsx
@@ -1,7 +1,0 @@
-
-function Dashboard() {
-    return (
-        <Dashboard/>
-    );
-}
-export default Dashboard;
