--- Lab_week_07/tailwind.config.js
@@ -1,19 +1,0 @@
-// tailwind.config.js
-/** @type {import('tailwindcss').Config} */
-export default {
-    content: ['./index.html', './src/**/*.{js,ts,jsx,tsx}'],
-    theme: {
-        extend: {
-            keyframes: {
-                slideUp: {
-                    '0%': { opacity: '0', transform: 'translateY(40px)' },
-                    '100%': { opacity: '1', transform: 'translateY(0)' },
-                },
-            },
-            animation: {
-                slideUp: 'slideUp 0.4s ease-out forwards',
-            },
-        },
-    },
-    plugins: [],
-};
